requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.1.1
pydantic==2.9.2
opencv-python==4.9.0.80
Pillow==10.4.0
PyMuPDF==1.24.8
yt-dlp==2024.11.18
